Updated After batch file creation
-No longer overwrite existing after.bat file. if it exists i only append the lines to the file otherwise if it doesnt exists i still create it

+++ artt_after.py
@@ -31,32 +31,47 @@ def createfile():
                 if '/setup/' in f:
                     dirs.remove(f)
 
-
     # for dirName, subdirList, fileList in os.walk(tc_base_dir):
     #     if 'setup' in dirName:
     #         dirs.append(dirName)
     #         for f in dirs:
     #             if '/setup/' in f:
     #                 dirs.remove(f)
-
     for d in dirs:
-        if '.DIR' in d.upper():
-            open(d + after_batch, 'w').close()
-            with open(d + after_batch, 'rw+') as a:
-                a.truncate(0)
-
-        try:
+        if os.path.exists(d + after_batch):
+            try:
+                if '.DIR' in d.upper():
+                    testcase = d.upper()
+                    testcase = testcase.rstrip('./DIR/SETUP')
+                    testcase = testcase.lstrip('C:/REGRESSN/CASES')
+                    with open(d + after_batch, 'a') as a:
+                        a.write('\n')
+                        a.write('REM File has been edited by artt_after.py\n')
+                        a.write('REM for use with customer ' + customer + ' test case ' + testcase + '\n')
+                        a.write(after_check + customer + ' ' + testcase + '\n')
+                        a.write(after_run + customer + ' ' + testcase + '\n')
+                        # shutil.copy(default_after, d + after_batch)
+            except:
+                pass
+        elif not os.path.exists(d + after_batch):
             if '.DIR' in d.upper():
-                testcase = d.upper()
-                testcase = testcase.rstrip('./DIR/SETUP')
-                testcase = testcase.lstrip('C:/REGRESSN/CASES')
+                open(d + after_batch, 'w').close()
                 with open(d + after_batch, 'rw+') as a:
-                    a.writelines('REM File has been edited by artt_after.py\n')
-                    a.writelines('REM for use with customer ' + customer + ' test case ' + testcase + '\n')
-                    a.writelines(after_check + customer + ' ' + testcase + '\n')
-                    a.writelines(after_run + customer + ' ' + testcase + '\n')
-                    # shutil.copy(default_after, d + after_batch)
-        except:
+                    a.truncate(0)
+            try:
+                if '.DIR' in d.upper():
+                    testcase = d.upper()
+                    testcase = testcase.rstrip('./DIR/SETUP')
+                    testcase = testcase.lstrip('C:/REGRESSN/CASES')
+                    with open(d + after_batch, 'rw+') as a:
+                        a.writelines('REM File has been edited by artt_after.py\n')
+                        a.writelines('REM for use with customer ' + customer + ' test case ' + testcase + '\n')
+                        a.writelines(after_check + customer + ' ' + testcase + '\n')
+                        a.writelines(after_run + customer + ' ' + testcase + '\n')
+                        # shutil.copy(default_after, d + after_batch)
+            except:
+                pass
+        else:
             pass
 
     print("All test cases updated with After.bat")
